# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Planificacion Horaria",
        "description": "Gestion de horarios de clases",
        "contact": {
            "email": "contacto@tudominio.com"
        },
        "license": {
            "name": "Licencia",
            "url": "https://www.terciarioUrquiza.com.ar/planificacionHoraria"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/comisiones": {
            "get": {
                "tags": [
                    "Comision"
                ],
                "summary": "Obtener todas las comisiones",
                "description": "Obtener todas las comisiones",
                "operationId": "97297c51f29bb66399681d7324fa959a",
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Comision"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener las comisiones"
                    }
                }
            }
        },
        "/api/comisiones/carreras/{carrera}": {
            "get": {
                "tags": [
                    "Comision"
                ],
                "summary": "Obtener comisiones por carrera",
                "description": "Obtener comisiones por carrera",
                "operationId": "0b3243c3ffd7cf5daa71d626fa41ae0d",
                "parameters": [
                    {
                        "name": "carrera",
                        "in": "path",
                        "description": "Carrera de la comision",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Comision"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener las comisiones"
                    }
                }
            }
        },
        "/api/comisiones/{id}": {
            "get": {
                "tags": [
                    "Comision"
                ],
                "summary": "Obtener comision por id",
                "description": "Obtener comision por id",
                "operationId": "getComisionPorId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de la comision",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comision"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Comision no encontrada"
                    },
                    "500": {
                        "description": "Error al obtener la comision"
                    }
                }
            }
        },
        "/api/comisiones/guardar": {
            "post": {
                "tags": [
                    "Comision"
                ],
                "summary": "Guardar comision",
                "description": "Guardar comision",
                "operationId": "4d6c475bbf4e91043728105696b3b700",
                "requestBody": {
                    "description": "Datos de la comision",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ComisionData"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ComisionData"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al guardar la comision"
                    }
                }
            }
        },
        "/api/comisiones/actualizar/{id}": {
            "put": {
                "tags": [
                    "Comision"
                ],
                "summary": "Actualizar comision",
                "description": "Actualizar comision",
                "operationId": "79d09493fa6d5acaad346b7984f4c58e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de la comision",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos de la comision",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ComisionData"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ComisionData"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Comision no encontrada"
                    },
                    "500": {
                        "description": "Error al actualizar la comision"
                    }
                }
            }
        },
        "/api/comisiones/eliminar/{id}": {
            "delete": {
                "tags": [
                    "Comision"
                ],
                "summary": "Eliminar comision",
                "description": "Eliminar comision",
                "operationId": "6e16fe9c8fc04739512c7156d7ef7723",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de la comision",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "404": {
                        "description": "Comision no encontrada"
                    },
                    "500": {
                        "description": "Error al eliminar la comision"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ComisionData": {
                "title": "ComisionData",
                "description": "Esquema del objeto ComisionData",
                "properties": {
                    "anio": {
                        "description": "Año de la comision",
                        "type": "integer",
                        "example": "2"
                    },
                    "division": {
                        "description": "Division de la comision",
                        "type": "Integer",
                        "example": "3"
                    },
                    "carrera": {
                        "description": "Carrera de la comision",
                        "type": "enum",
                        "example": "DS"
                    }
                },
                "type": "object"
            },
...
            "Comision": {
                "title": "Comision",
                "description": "Esquema del objeto Comision",
                "properties": {
                    "anio": {
                        "description": "Año de la comision",
                        "type": "integer"
                    },
                    "division": {
                        "description": "Division de la comision",
                        "type": "string"
                    },
                    "carrera": {
                        "description": "Carrera de la comision",
                        "type": "string"
                    },
                    "fecha_creacion": {
                        "description": "Fecha de creacion de la comision",
                        "type": "string",
                        "format": "date-time"
                    },
                    "fecha_modificacion": {
                        "description": "Fecha de modificacion de la comision",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    }
}